use the given Galera resource name when checking for existing service
The code here creates a new Service with Galera.Name, but
when it checks for that name already existing, it was hardcoded
to "openstack"; updated this to check for the Galera.Name
that will ultimately be used.   Previously, the controller would
continuously try to create the Service each time as the check
would fail, leading it to never complete the Reconcile process.
This issue was identified as I have been creating Galera
instances with custom names in a custom namespace.